Further simplifications
